requirements.txt
django==2.2.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-filter==2.1.0
django-cors-headers==3.2.1
coreapi==2.3.3
coreapi-cli==1.0.9
requests==2.20.0
django-extensions==2.1.2
django-debug-toolbar==1.11
django-rest-swagger==2.2.0
django-debug-toolbar==1.11